requirements/base.txt
bs4
ccxt
cryptocompare==0.6
cryptography==2.3
celery==4.4.0
django-celery-beat==1.1.1
django==2.2.20
django-csp==3.7
django-cors-headers==2.4.0
django-filter==2.0.0
django-haystack
django-ratelimit==1.1.0
djangorestframework==3.11.2
gitterpy
gunicorn
matplotlib
Markdown==3.2.1
reportlab==3.5.6
requests
requests_oauthlib
Pillow==4.0.0
premailer
populus
psycopg2-binary==2.7.5
PyJWT==1.5.3
PyPdf2
python-twitter==3.2
sendgrid==5.6.0
slackclient==2.0.0
eth-tester
elasticsearch==7.10.0
websockets
web3==4.5.0
eth-abi==1.1.1
mailchimp3==3.0.7
django_chartit==0.2.9
MechanicalSoup==0.9.0.post4
pyshorteners==0.6.1
statistics
django-email-obfuscator
cytoolz==0.9.0
boto==2.49.0
google-api-python-client==1.7.3
django-environ==0.4.5
eth-utils==1.4.1
jsondiff==1.1.1
social-auth-app-django==4.0.0
django-ipware==2.0.2
geoip2==2.8.0
django-silk==2.0.0
django-extensions==2.2.1
ipdb==0.11
django-autotranslate==1.1.1
svgutils==0.3.0
watchdog==0.9.0
Werkzeug[watchdog]==0.15.5
imageio
boto3==1.7.81
django-storages==1.11.1
easy-thumbnails==2.5
eth-account==0.2.2
django-classy-tags==0.8.0
django-cookie-law==2.0.1
django-impersonate==1.3
pg_activity
PyYAML==4.2b4
ecdsa
pysha3
oyaml
pygithub==1.54.0
git+https://github.com/gitcoinco/django-cacheops.git#egg=django-cacheops
django-redis==4.10.0
collectfast==0.6.2
django-health-check==3.7.0
elastic-apm==3.0.1
mkdocs==1.0.4
pymdown-extensions==5.0
mkdocs-material==3.0.4
pydoc-markdown==2.0.4
oauth2client==4.1.3
pyvips==2.1.4
redis-semaphore
watchtower==0.5.4
Wand==0.4.4
raven==6.9.0
sentry-sdk==0.19.5
websocket-client
bleach
python-magic
django-oauth-toolkit
ddtrace
metadata-parser==0.10.0
redis==3.3.11
wiki==0.7.3
django-bulk-update
pdfrw
django-admin-sortable2==0.7.6
twilio
django-queryset-csv
django-proxy==1.2.1
ed25519
tweepy
attrs==19.3.0
oogway==0.7.0
icalendar==4.0.7
duniterpy==0.61.0
aiohttp >= 3.6.3
jsonschema==3.2.0
pyPEG2==2.15.2
base58==2.1.0
libnacl==1.7.2
pyaes==1.6.1
numpy==1.19.5
rjsmin==1.1.0
rcssmin==1.0.6
libsass==0.20.1
graphqlclient==0.2.4

channels==2.4.0
channels_redis
oogway==0.7.0
daphne==2.4.0
